Use factory function to generate rest api
